Rewrite how title is extracted
Previously, it used the indexTitle field from the instance record, but
that's wrong because the index title omits words like "the".  The
"title" field in the record needs to be parsed to separate the title
and author, so it's more annoying to use, but that's what the new code
does.