Go back to getting authors from the title field
The `contributors` list field in the records sometimes only contains
the first author and no indication that there may be other authors.
(The `primary` field of the author entry in the contributors list is
not consistently `False` for multiple authors, making it impossible to
tell whether there are more than one author.)  So, I've gone back to
parsing the title text to get better author data.